+++ easypost/referral.py
@@ -15,6 +15,7 @@ from easypost.requestor import (
 )
 
 
+# TODO: Rename `Referral` to `ReferralCustomer` and have the module name match as well
 class Referral:
     @staticmethod
     def create(
@@ -94,7 +95,11 @@ class Referral:
 
         try:
             stripe_token = Referral._create_stripe_token(
-                number, expiration_month, expiration_year, cvc, easypost_stripe_api_key
+                number,
+                expiration_month,
+                expiration_year,
+                cvc,
+                easypost_stripe_api_key,
             )
         except Exception:
             raise Error(message="Could not send card details to Stripe, please try again later")
@@ -105,6 +110,67 @@ class Referral:
             priority=primary_or_secondary,
         )
         return convert_to_easypost_object(response)
+
+    @staticmethod
+    def add_payment_method(
+        stripe_customer_id: str,
+        payment_method_reference: str,
+        priority: Optional[str] = "primary",
+    ) -> Dict[str, Any]:
+        """Add a Stripe payment method to your EasyPost account.
+
+        This endpoint uses a user's personal Stripe account. The `stripe_customer_id`
+        and `payment_method_reference` IDs both come from Stripe. By adding these to
+        EasyPost, we will associate your Stripe payment method with either your primary
+        or secondary EasyPost payment method.
+        """
+        requestor = Requestor()
+        wrapped_params = {
+            "payment_method": {
+                "stripe_customer_id": stripe_customer_id,
+                "payment_method_reference": payment_method_reference,
+                "priority": priority,
+            }
+        }
+
+        response, api_key = requestor.request(
+            method=RequestMethod.POST,
+            url="/referral_customers/payment_method",
+            params=wrapped_params,
+            beta=True,
+        )
+
+        return convert_to_easypost_object(response=response, api_key=api_key)
+
+    @staticmethod
+    def refund_by_amount(refund_amount: str) -> Dict[str, Any]:
+        """Refund a ReferralCustomer wallet by specifying an amount."""
+        requestor = Requestor()
+        wrapped_params = {"refund_amount": refund_amount}
+
+        response, api_key = requestor.request(
+            method=RequestMethod.POST,
+            url="/referral_customers/refunds",
+            params=wrapped_params,
+            beta=True,
+        )
+
+        return convert_to_easypost_object(response=response, api_key=api_key)
+
+    @staticmethod
+    def refund_by_payment_log(payment_log_id: str) -> Dict[str, Any]:
+        """Refund a ReferralCustomer wallet by specifying a payment log ID to completely refund."""
+        requestor = Requestor()
+        wrapped_params = {"payment_log_id": payment_log_id}
+
+        response, api_key = requestor.request(
+            method=RequestMethod.POST,
+            url="/referral_customers/refunds",
+            params=wrapped_params,
+            beta=True,
+        )
+
+        return convert_to_easypost_object(response=response, api_key=api_key)
 
     @staticmethod
     def _retrieve_easypost_stripe_api_key() -> str: