+++ tests/test_referral.py
@@ -9,6 +9,7 @@ import easypost
 REFERRAL_USER_PROD_API_KEY = os.getenv("REFERRAL_USER_PROD_API_KEY", "123")
 
 
+# TODO: Rename this file to `test_referral_customer.py` and all test names need `customer` added to them
 @pytest.mark.vcr()
 def test_referral_user_create(partner_prod_api_key):
     """This test requires a partner user's production API key via PARTNER_USER_PROD_API_KEY."""
@@ -89,3 +90,43 @@ def test_referral_add_credit_card_error(mock_stripe_token, mock_easypost_key, cr
         )
 
     assert str(error.value) == "Could not send card details to Stripe, please try again later"
+
+
+@pytest.mark.vcr()
+def test_referral_add_payment_method(referral_customer_prod_api_key):
+    """This test requires a referral customer user's production API key via REFERRAL_CUSTOMER_PROD_API_KEY.
+
+    We expect this test to fail because we don't have valid Stripe details to use. Assert the correct error.
+    """
+    with pytest.raises(Exception) as error:
+        easypost.Referral.add_payment_method(
+            stripe_customer_id="cus_123",
+            payment_method_reference="ba_123",
+            priority="primary",
+        )
+
+    assert str(error.value) == "Invalid Payment Gateway Reference."
+
+
+@pytest.mark.vcr()
+def test_referral_refund_by_amount(referral_customer_prod_api_key):
+    """This test requires a referral customer user's production API key via REFERRAL_CUSTOMER_PROD_API_KEY.
+
+    We expect this test to fail because we don't have valid billing details to use. Assert the correct error.
+    """
+    with pytest.raises(Exception) as error:
+        easypost.Referral.refund_by_amount(refund_amount="$20.00")
+
+    assert str(error.value) == "Refund amount doesn't match. Please escalate this to finance, this is unexpected."
+
+
+@pytest.mark.vcr()
+def test_referral_refund_by_payment_log(referral_customer_prod_api_key):
+    """This test requires a referral customer user's production API key via REFERRAL_CUSTOMER_PROD_API_KEY.
+
+    We expect this test to fail because we don't have valid billing details to use. Assert the correct error.
+    """
+    with pytest.raises(Exception) as error:
+        easypost.Referral.refund_by_payment_log(payment_log_id="paylog_123")
+
+    assert str(error.value) == "We could not find a transaction with that id."